feat: server action 추상화
- 모임 관련 함수는 co-location관점으로 분류
- 각각의 파일 분리보다는 한 파일내부에서 다룸
- 삭제의 경우 유효성 검사 불필요
